Clear up the asteroid processing pipeline
